docs(site): separate local server API reference
